Configured ExportToTerra, has initialization error when accessed directly #465
