+++ explorer/app/components/Export/components/ExportToTerra/exportToTerra.stories.tsx
@@ -3,11 +3,17 @@ import React from "react";
 import { ExportToTerra } from "./exportToTerra";
 
 export default {
+  argTypes: {
+    path: { control: "text" },
+    url: { control: "text" },
+  },
   component: ExportToTerra,
   title: "Components/Section/Export/ExportToTerra",
 } as ComponentMeta<typeof ExportToTerra>;
 
-const Template: ComponentStory<typeof ExportToTerra> = () => <ExportToTerra />;
+const Template: ComponentStory<typeof ExportToTerra> = (args) => (
+  <ExportToTerra {...args} />
+);
 
 export const NotStartedExportToTerra = Template.bind({});
 NotStartedExportToTerra.args = {};
